Improve custom quasi-quotation parsing
Problem: currently parsing `[int||Text|]` fails with the current Haskell
grammar.

This happens, because we parse the beginning in a greedy way, and so
parse `[int||` as beginning of typed quasi-quoter. Thus end must be
`||]` respectively, but we never meet it.

Solution: custom quoters are never typed, and GHC (tried with 8.10.7
version) interprets `[int||Abc||]` as a mere quoter with `|Abc|` body.

So here we adjust the pattern for custom quoters to be never typed.